Refactor: Extract operators to separate module (Phase 2.1)
Moves operator enums from ast/lib.rs to dedicated operators.rs module,
improving code organization and establishing pattern for AST refactoring.

Changes:
- Created crates/ast/src/operators.rs (39 lines)
  - BinaryOperator enum (13 variants)
  - UnaryOperator enum (5 variants)
- Updated crates/ast/src/lib.rs
  - Added mod operators declaration
  - Added public re-exports
  - Removed old operator definitions (39 lines)
  - File reduced from 566 to 527 lines

Verification:
✅ All 27 AST tests pass
✅ No clippy warnings
✅ operators.rs is ~40 lines (meets acceptance criteria)
✅ lib.rs reduced by 39 lines as expected

This is the first and safest module extraction - operators have zero
dependencies on other AST types, making this a clean separation.

Closes #28
